backend/requirements.txt
flask
flask-cors
pymongo
pyjwt
python-dotenv
werkzeug